# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rust Spell Server",
    "version": "0.1.0",
    "description": "A production-ready Rust HTTP server exposing a spell-checking API over OpenAPI-compliant endpoints.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "contact": {
      "name": "Brian Tafoya",
      "email": "btafoya@briantafoya.com",
      "url": "https://github.com/btafoya/rustspell-server"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [],
  "paths": {
...
    "/spellcheck": {
      "post": {
        "operationId": "spellcheck",
        "tags": ["Spell-check"],
        "summary": "Spell-check tokens",
        "description": "Accepts text and/or an explicit word list and returns one result per token occurrence.",
        "security": [{ "apiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SpellCheckRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spell-check results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SpellCheckResponse" }
              }
            }
          },
          "400": {
            "description": "Validation error (neither `text` nor `words` provided, or malformed `language`), or `language` is a well-formed but unloadable code (download/parse failure).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, or revoked key",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": {
            "description": "Rate limited or over quota. Two distinct causes share this status: too many failed authentication attempts from this IP (`type` ends in `rate-limited`, retries after `Retry-After` seconds), or the tenant's request quota is exhausted for the current billing period (`type` ends in `quota-exceeded`, no `Retry-After` — resolved by the platform/billing app via `PATCH /tenants/{id}`, not by waiting).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SpellCheckRequest": {
        "type": "object",
        "description": "Request body for spell-check endpoints. At least one of `text` or `words` must be provided.",
        "minProperties": 1,
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "description": "Free-form text to tokenize and spell-check"
          },
          "words": {
            "type": "array",
            "maxItems": 1000,
            "items": { "type": "string" },
            "description": "Explicit list of words to spell-check"
          },
          "language": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,20}$",
            "description": "Overrides the calling tenant's default language for this request. Downloaded and cached on first use if not already loaded; an unloadable language returns 400, not 500."
          }
        }
      },
      "SpellCheckResponse": {
        "type": "object",
        "required": ["results"],
        "properties": {
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TokenResult" }
          }
        }
      },
      "TokenResult": {
        "type": "object",
        "required": ["token", "valid", "suggestions"],
        "properties": {
          "token": { "type": "string" },
          "valid": { "type": "boolean" },
          "suggestions": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Suggested corrections. Empty when `valid` is `true`."
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}